Fix Windows build
