+++ soapcalls/utils/substools.go
@@ -11,7 +11,6 @@ import (
 	"path/filepath"
 	"strconv"
 
-	"github.com/alexballas/chardet"
 	"github.com/mitchellh/mapstructure"
 )
 
@@ -119,24 +118,4 @@ func ExtractSub(ffmpeg string, n int, f string) (string, error) {
 	}
 
 	return tempSub.Name(), nil
-}
-
-func getCharDet(path string) (string, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return "", err
-	}
-	defer f.Close()
-
-	firstBytes := make([]byte, 512)
-
-	n, err := f.Read(firstBytes)
-	if err != nil {
-		return "", err
-	}
-
-	det := chardet.NewTextDetector()
-	charGuess, err := det.DetectBest(firstBytes[:n])
-
-	return charGuess.Charset, err
 }
